Meca Mind mount tests
