+++ src/test/java/br/unicamp/meca/mind/action/Test1ActionFromPlanningCodelet.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package br.unicamp.meca.mind;
+package br.unicamp.meca.mind.action;
 
 import java.util.ArrayList;
 
@@ -13,9 +13,9 @@ import br.unicamp.meca.system1.codelets.ActionFromPlanningCodelet;
  * @author andre
  *
  */
-public class Test2ActionFromPlanningCodelets extends ActionFromPlanningCodelet {
+public class Test1ActionFromPlanningCodelet extends ActionFromPlanningCodelet {
 
-	public Test2ActionFromPlanningCodelets(String id, ArrayList<String> perceptualCodeletsIds, String motorCodeletId,
+	public Test1ActionFromPlanningCodelet(String id, ArrayList<String> perceptualCodeletsIds, String motorCodeletId,
 			String soarCodeletId) {
 		super(id, perceptualCodeletsIds, motorCodeletId, soarCodeletId);
 	}
@@ -32,10 +32,8 @@ public class Test2ActionFromPlanningCodelets extends ActionFromPlanningCodelet {
 			if(memory.getI()!=null && memory.getI() instanceof String) {
 				String perceptualContent = (String) memory.getI();
 				
-				((MemoryContainer) motorMemory).setI("Test2ActionFromPlanning - "+perceptualContent,getActivation(),id);
+				((MemoryContainer) motorMemory).setI("Test1ActionFromPlanning - "+perceptualContent,getActivation(),id);
 			}
 		}
-
 	}
-
 }
