+++ src/test/java/br/unicamp/meca/mind/action/Test2ActionFromPlanningCodelets.java
@@ -1,28 +1,27 @@
 /**
  * 
  */
-package br.unicamp.meca.mind;
+package br.unicamp.meca.mind.action;
 
 import java.util.ArrayList;
 
 import br.unicamp.cst.core.entities.Memory;
 import br.unicamp.cst.core.entities.MemoryContainer;
-import br.unicamp.meca.system1.codelets.ActionFromPerceptionCodelet;
+import br.unicamp.meca.system1.codelets.ActionFromPlanningCodelet;
 
 /**
  * @author andre
  *
  */
-public class Test1ActionFromPerceptionCodelet extends ActionFromPerceptionCodelet {
+public class Test2ActionFromPlanningCodelets extends ActionFromPlanningCodelet {
 
-	public Test1ActionFromPerceptionCodelet(String id, ArrayList<String> perceptualCodeletsIds,
-			ArrayList<String> motivationalCodeletsIds, String motorCodeletId, String soarCodeletId) {
-		super(id, perceptualCodeletsIds, motivationalCodeletsIds, motorCodeletId, soarCodeletId);
+	public Test2ActionFromPlanningCodelets(String id, ArrayList<String> perceptualCodeletsIds, String motorCodeletId,
+			String soarCodeletId) {
+		super(id, perceptualCodeletsIds, motorCodeletId, soarCodeletId);
 	}
 
 	@Override
 	public void proc(ArrayList<Memory> perceptualMemories, Memory broadcastMemory, Memory motorMemory) {
-		
 		if(perceptualMemories == null || perceptualMemories.size() == 0) {
 			return;
 		}
@@ -33,9 +32,10 @@ public class Test1ActionFromPerceptionCodelet extends ActionFromPerceptionCodele
 			if(memory.getI()!=null && memory.getI() instanceof String) {
 				String perceptualContent = (String) memory.getI();
 				
-				((MemoryContainer) motorMemory).setI("Test1ActionFromPerception - "+perceptualContent,getActivation(),id);
+				((MemoryContainer) motorMemory).setI("Test2ActionFromPlanning - "+perceptualContent,getActivation(),id);
 			}
 		}
+
 	}
 
 }
