+++ src/test/java/br/unicamp/meca/mind/perceptual/TestPerceptualCodelet.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package br.unicamp.meca.mind;
+package br.unicamp.meca.mind.perceptual;
 
 import java.util.ArrayList;
